requirements.txt
yfinance>=0.2.40
fredapi>=0.5.1
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
python-dotenv>=1.0
requests>=2.31
matplotlib>=3.7
pytest>=7.4
pytest-mock>=3.11
